+++ lab_5/main.py
